% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprolog.R
\name{pl_consult}
\alias{pl_consult}
\title{Consult a Prolog source file}
\usage{
pl_consult(pl, file)
}
\arguments{
\item{pl}{A \code{prolog_engine}.}

\item{file}{Path to a \code{.pl} file.}
}
\value{
\code{TRUE} if the file loaded without error.
}
\description{
Consult a Prolog source file
}
