% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprolog.R
\name{pl_consult_text}
\alias{pl_consult_text}
\title{Consult Prolog source held in a character string}
\usage{
pl_consult_text(pl, text)
}
\arguments{
\item{pl}{A \code{prolog_engine}.}

\item{text}{Prolog source text.}
}
\value{
\code{TRUE} if it loaded without error.
}
\description{
Consult Prolog source held in a character string
}
