% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprolog.R
\name{pl_eval}
\alias{pl_eval}
\title{Run a directive/goal via pl_eval}
\usage{
pl_eval(pl, expr)
}
\arguments{
\item{pl}{A \code{prolog_engine}.}

\item{expr}{A bare goal (no leading \code{:-} or trailing \code{.}).}
}
\value{
\code{TRUE} on success.
}
\description{
Run a directive/goal via pl_eval
}
